python/requirements.txt
web3~=6.13.0
hexbytes~=1.0.0
absl-py~=2.0.0
